+++ backend/backend-projeto-integrador/src/main/java/com/backendprojetointegrador/lajeDev/domain/service/CategoriaService.java
@@ -1,7 +1,7 @@
 package com.backendprojetointegrador.lajeDev.domain.service;
 
-import com.backendprojetointegrador.lajeDev.domain.exception.EntidadeNaoEncontrada;
 import com.backendprojetointegrador.lajeDev.domain.exception.RecursoJaExistenteException;
+import com.backendprojetointegrador.lajeDev.domain.exception.RecursoNaoEncontrado;
 import com.backendprojetointegrador.lajeDev.domain.model.Categoria;
 import com.backendprojetointegrador.lajeDev.domain.repository.ICategoriaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,8 +27,8 @@ public class CategoriaService {
 
     public Categoria buscarCategoria(Long idCategoria) {
         return categoriasRepository.findById(idCategoria)
-                .orElseThrow(() -> new EntidadeNaoEncontrada("Categoria com de id: " + idCategoria
-                        + " não encontrada ou inexistente!"));
+                .orElseThrow(() -> new RecursoNaoEncontrado("Categoria com de id: " + idCategoria
+                        + " não existe!"));
     }
 
     public List<Categoria> listarCategorias() {
@@ -36,6 +36,10 @@ public class CategoriaService {
     }
 
     public void excluirCategoria(Long idCategoria) {
+        if (!existeCategoria(idCategoria)) {
+            throw new RecursoNaoEncontrado("Categoria com de id: " + idCategoria
+                    + " não existe!");
+        }
         categoriasRepository.deleteById(idCategoria);
     }
